requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
Pillow==10.0.0
phonenumbers==8.13.13
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
